+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/cli
 
+## 0.5.112
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.50
+    - @stoker-platform/node-client@0.5.71
+    - @stoker-platform/utils@0.5.62
+
 ## 0.5.111
 
 ### Patch Changes