+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.71
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.50
+    - @stoker-platform/utils@0.5.62
+
 ## 0.5.70
 
 ### Patch Changes